recognition_multi, gui: Add some custom class for gui and improve support for multi threading
Add package gui that contains two files: MainWindow.py and Frame.py.
MainWindow.py contains a class "MainWindow" that is a custom class for tk.Tk. At the moment, it isn't useful, but maybe later it can be improved.
Frame.py contains 3 classes, one for each frame in main window. At the moment, it's useful too, but maybe later it can be.

For multi threading, added support to print in the log box partial result in the right order, and some other fixes that I don't remember.