didspeech, parsing, prints: Add loading percents. Add "all format" in select file
Now it prints a percents to follow loading process.
Now when you select "select file" there is "All format" option.
PrintPartial and PrintLoading stop when reach 100%.
Percents is chunks_done/total_chunks*100. These variables are in Didspeech class as class variables. PrintPartial and PrintLoading (as other class) have _parent variable as instance variable. It is the main Didspeech variable (so, the app). So they can access to his fields.